mail-ai-backend/services/auth_service/requirement.txt
fastapi
uvicorn
motor
google-auth-oauthlib
google-api-python-client
python-dotenv
